requirements.txt
Twisted==14.0.0
Werkzeug==0.9.6
cffi==0.8.6
characteristic==0.1.0
cryptography==0.5.2
klein==0.2.3
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
service-identity==1.0.0
six==1.7.3
treq==0.2.1
wsgiref==0.1.2
zope.interface==4.1.1
